docker-image/apiserver/requirements.txt
aniso8601==9.0.1
appdirs==1.4.4
attrs==20.3.0
click==7.1.2
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-RESTful==0.3.8
flask-restx==0.2.0
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==1.1.1
mysql-connector-python==8.0.24
mysqlclient==2.0.3
prometheus-client==0.9.0
prometheus-flask-exporter==0.18.1
protobuf==3.15.8
pyrsistent==0.17.3
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
virtualenv==20.4.2
Werkzeug==1.0.1
